requirements.txt
requests>=2.31.0
python-wordpress-xmlrpc>=2.3
qiniu>=7.13.0
python-dateutil>=2.8.2